Struct smithy_xml::encode::XmlWriter

pub struct XmlWriter<'a> { /* fields omitted */ }

XmlWriter Abstraction

XmlWriter (and friends) make generating an invalid XML document a type error. Nested branches of the Xml document mutable borrow from the root. You cannot continue writing to the root until the nested branch is dropped and dropping the nested branch writes the terminator (eg. closing element).

The one exception to this rule is names—it is possible to construct an invalid Xml Name. However, names are always known ahead of time and always static, so this would be obvious from the code.

Furthermore, once const panic stabilizes, we’ll be able to make an invalid XmlName a compiler error.

Example

use smithy_xml::encode::XmlWriter;
let mut s = String::new();
let mut doc = XmlWriter::new(&mut s);
let mut start_el = doc.start_el("Root")
    .write_ns("http://example.com", None);
let mut start_tag = start_el.finish();
start_tag.data("hello");
start_tag.finish();
assert_eq!(s, "<Root xmlns=\"http://example.com\">hello</Root>");

See tests/handwritten_serializers.rs for more usage examples.

Implementations

impl<'a> XmlWriter<'a>

pub fn new(doc: &'a mut String) -> Self

impl<'a> XmlWriter<'a>

pub fn start_el<'b, 'c>(&'c mut self, tag: &'b str) -> ElWriter<'c, 'b>